BRD-MergedTenantMaster-Backend/requirements.txt
boto3==1.29.7
celery==5.3.4
django==4.2.7
django-axes
django-axes>=6.3.0
django-cors-headers==4.2.0
django-filter
django-filter==23.5
django-guardian==2.4.0
django-multiselectfield==0.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
django-tenant-schemas==1.11.0
gunicorn==21.2.0
pyotp==2.9.0
python-decouple==3.8
qrcode==8.2
redis==5.0.1
requests==2.31.0
setuptools<81
whitenoise==6.6.0
